[❌FIX] registro de bodega monto de tipo de pago, reubicación método de políticas
