+++ sigebo/src/main/java/com/sicobo/sicobo/controller/user/UserController.java
@@ -1,13 +1,9 @@
 package com.sicobo.sicobo.controller.user;
 
-import com.sicobo.sicobo.dto.DTOPolicies;
-import com.sicobo.sicobo.model.BeanPolicies;
-import com.sicobo.sicobo.serviceimpl.PoliciesServiceImpl;
 import com.sicobo.sicobo.serviceimpl.WarehouseServiceImpl;
 import com.sicobo.sicobo.util.Message;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -15,7 +11,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
-import static com.sicobo.sicobo.util.Constantes.ObjectMessages.MESSAGE_CATCH_ERROR;
 import static com.sicobo.sicobo.util.Constantes.Redirects.*;
 import static com.sicobo.sicobo.util.Constantes.Roles.ROLE_USUARIO;
 import static com.sicobo.sicobo.util.Constantes.Stuff.*;
@@ -23,36 +18,11 @@ import static com.sicobo.sicobo.util.Constantes.Stuff.*;
 @Controller
 @Slf4j
 public class UserController {
-    private final PoliciesServiceImpl policiesService;
-
     private final WarehouseServiceImpl warehouseService;
 
     @Autowired
-    public UserController(PoliciesServiceImpl policiesService, WarehouseServiceImpl warehouseService) {
-        this.policiesService = policiesService;
+    public UserController(WarehouseServiceImpl warehouseService) {
         this.warehouseService = warehouseService;
-    }
-
-    @GetMapping("/politicas")
-    public String mostrarContenido(Model model, DTOPolicies termsAndConditions) {
-        try{
-            ResponseEntity<?> responseEntity = policiesService.listar();
-            Message message = (Message) responseEntity.getBody();
-            assert message != null;
-            BeanPolicies termsAndConditionsSearched = (BeanPolicies) message.getResult();
-
-            if (termsAndConditionsSearched.getDescription() != null){
-                termsAndConditions.setDescription(termsAndConditionsSearched.getDescription());
-            }else{
-                termsAndConditions.setDescription("No se han establecido los términos y condiciones");
-            }
-
-            model.addAttribute(TERMSANDCONDTIONS, termsAndConditions);
-        }catch (Exception e){
-            model.addAttribute(MESSAGE, MESSAGE_CATCH_ERROR);
-            log.error("Ocurrió un error en UserController - mostrarContenido" + e.getMessage());
-        }
-        return USER_TERMSANDCONDITIONS;
     }
 
     @Secured({ROLE_USUARIO})
